Get errors to vanish when window is closed
