+++ frontend/actions/session_actions.js
@@ -3,6 +3,7 @@ import * as APIUtil from '../util/session_api_util';
 export const RECEIVE_CURRENT_USER = 'RECEIVE_CURRENT_USER';
 export const RECEIVE_ERRORS = 'RECEIVE_ERRORS';
 export const REMOVE_CURRENT_USER = 'REMOVE_CURRENT_USER';
+export const REMOVE_ERRORS = 'REMOVE_ERRORS';
 
 export const receiveCurrentUser = (currentUser) => {
   return {
@@ -24,6 +25,14 @@ export const receiveErrors = (errors) => {
       errors: errors.responseJSON
   };
 };
+
+export const removeErrors = (errors) => {
+  return {
+    type: REMOVE_ERRORS,
+    errors
+  };
+};
+
 
 export const signup = user => dispatch => {
   return APIUtil.signup(user)
