+++ frontend/components/session_form_container.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { connect } from 'react-redux';
 import SessionForm from './session_form';
 import zzSessionForm from './zzsession_form';
-import { login, logout, signup } from '../actions/session_actions';
+import { login, logout, signup, removeErrors } from '../actions/session_actions';
 
 const mapStateToProps = ({ session }) => {
   return {
@@ -15,7 +15,8 @@ const mapDispatchToProps = (dispatch) => {
   return {
     logout: () => dispatch(logout()),
     login: (user) => dispatch(login(user)),
-    signup: (user) => dispatch(signup(user))
+    signup: (user) => dispatch(signup(user)),
+    removeErrors: () => dispatch(removeErrors())
   };
 };
