+++ frontend/components/some_form.jsx
@@ -13,7 +13,6 @@ class SomeForm extends React.Component {
   }
 
   handleSubmit(e) {
-    debugger
     e.preventDefault();
     const user = Object.assign({}, this.state);
     this.props.processForm(user).then(() => this.props.closeModal());
