+++ frontend/reducers/session_reducer.js
@@ -1,4 +1,4 @@
-import { RECEIVE_ERRORS, RECEIVE_CURRENT_USER, REMOVE_CURRENT_USER } from '../actions/session_actions';
+import { RECEIVE_ERRORS, RECEIVE_CURRENT_USER, REMOVE_CURRENT_USER, REMOVE_ERRORS } from '../actions/session_actions';
 import merge from 'lodash/merge';
 
 const defaultState = {
@@ -12,9 +12,9 @@ const sessionReducer = (state = defaultState, action) => {
       const safeUser = (({id, username }) => ({ id, username }))(action.currentUser);
       return merge({}, state, { currentUser: safeUser });
     case RECEIVE_ERRORS:
-    debugger
-      state.errors = [];
-      return merge({}, state, { errors: action.errors });
+      return Object.assign({}, state, { errors: action.errors });
+    case REMOVE_ERRORS:
+      return Object.assign({}, state, { errors: []});
     case REMOVE_CURRENT_USER:
       return defaultState;
     default:
